+++ frontend/src/pages/MyList.js
@@ -1,12 +1,12 @@
 import React from "react";
 // Connecting the NavBar to the pages
+import FooterComponent from "../components/FooterComponent";
 import Navbar from "../components/NavBar";
-
 export default function MyList() {
   return (
-    <><Navbar /><div> 
+    <><><Navbar /><div>
       <h1>My List page</h1>
-    </div></>
+    </div></><FooterComponent /></>
 
   );
 }