+++ frontend/src/pages/Support.js
@@ -1,12 +1,13 @@
 import React from "react";
 // Connecting the NavBar to the pages
+import FooterComponent from "../components/FooterComponent";
 import Navbar from "../components/NavBar";
 
 export default function Support() {
   return (
-    <><Navbar /><div> 
+    <><><Navbar /><div>
       <h1>Customer Support</h1>
-    </div></>
+    </div></><FooterComponent /></>
 
   );
 }