+++ frontend/src/pages/SurpriseMe.js
@@ -1,12 +1,13 @@
 import React from "react";
 // Connecting the NavBar to the pages
+import FooterComponent from "../components/FooterComponent";
 import Navbar from "../components/NavBar";
 
 export default function SurpriseMe() {
   return (
-    <><Navbar /><div>
+    <><><Navbar /><div>
       <h1>Surprise Me page</h1>
-    </div></>
+    </div></><FooterComponent /></>
   );
 }